refactored block_kit_builder and renamed directory blocks
