+++ flameboi/blocks/layout_block/divider_block.py
@@ -1,4 +1,4 @@
-from .layout_block import LayoutBlock
+from flameboi.blocks.layout_block.layout_block import LayoutBlock
 
 
 class DividerBlock(LayoutBlock):
